Moved film into the camera class

+++ src/common.hpp
@@ -15,8 +15,7 @@ struct boundary
     spacial_t max;
 };
 
-// We use plain-old floats for colour - we will never need double precision for this
-using colour_t = math::vector<float, 3>;
+using colour_t = math::vector<double, 3>;
 
 // The actual RBG values of the pixels. We template for optional 8-bit / 16-bit colour
 template <typename T = std::uint8_t>
@@ -26,12 +25,12 @@ using rgb_t = math::vector<T, 3>;
 using rgb8_t  = rgb_t<std::uint8_t>;
 using rgb16_t = rgb_t<std::uint16_t>;
 
-// Converts a floating point colour to its corresponding RGB value. Floating point values over 1 are truncated
+// Converts a floating point colour to its corresponding RGB value (values over 1 are truncated)
 template <typename T = std::uint8_t>
 constexpr rgb_t<T> to_rgb(const colour_t& colour)
 {
     rgb_t<T> ret;
     for (std::size_t i = 0; i < 3; i++)
-        ret[i] = std::round(std::min(std::abs(colour[i]), 1.0f) * std::numeric_limits<T>::max());
+        ret[i] = std::round(std::min(std::abs(colour[i]), 1.0) * std::numeric_limits<T>::max());
     return ret;
 }
